Teach CI to run further checks
Also bump OTP in CI for benefiting from the
[fix](https://github.com/erlang/otp/commit/304fef65a1#diff-77319525f6e8aa302e3bb8424802a2a3)
of a bug present in Dialyzer of OTP 19.3.

The command `make unit-tests` attempts to run both eunit tests (none)
and ct tests (none).  Attempt to run ct tests fails with
`no_dir_specified`.  Ignore failure anyways as there are no unit tests
at the moment, and the approach for writing unit tests is not planted
yet.